Task 4
NumPy library provides the mathematical operation at once on the array for computation, it is Used for linear algebra, random number generation , and many more. Numpy provides support for large multi-dimensional array and matrices, which are important for scientific computation. It is faster than did the operation on the Python list. It also works with libraries such as Pandas, Matplotlib , Scipy, etc